+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
+- Corrected reporting on the configurabilty of the PropertyDescriptor (configurable: true)
+
 ### Added
 ### Changed
 ### Removed